Rename header to function dispatch
